refactor(jmxlogger): migrate comp/agent/jmxlogger to V2 component architecture
