+++ comp/agent/jmxlogger/BUILD.bazel
@@ -5,4 +5,7 @@ go_library(
     srcs = ["component.go"],
     importpath = "github.com/DataDog/datadog-agent/comp/agent/jmxlogger",
     visibility = ["//visibility:public"],
+    deps = [
+        "//comp/agent/jmxlogger/def",
+    ],
 )
