+++ comp/agent/jmxlogger/def/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "def",
+    srcs = ["component.go"],
+    importpath = "github.com/DataDog/datadog-agent/comp/agent/jmxlogger/def",
+    visibility = ["//visibility:public"],
+)
