+++ comp/agent/jmxlogger/fx/fx.go
@@ -1,0 +1,24 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+// Package fx provides the fx module for the jmxlogger component.
+package fx
+
+import (
+	"go.uber.org/fx"
+
+	jmxloggerimpl "github.com/DataDog/datadog-agent/comp/agent/jmxlogger/impl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+func Module(params jmxloggerimpl.Params) fxutil.Module {
+	return fxutil.Component(
+		fxutil.ProvideComponentConstructor(
+			jmxloggerimpl.NewComponent,
+		),
+		fx.Supply(params),
+	)
+}
