+++ comp/agent/jmxlogger/impl/BUILD.bazel
@@ -1,0 +1,32 @@
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "impl",
+    srcs = [
+        "jmxlogger.go",
+        "params.go",
+    ],
+    importpath = "github.com/DataDog/datadog-agent/comp/agent/jmxlogger/impl",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//comp/agent/jmxlogger/def",
+        "//comp/core/config",
+        "//comp/def",
+        "//pkg/util/defaultpaths",
+        "//pkg/util/log/setup",
+    ],
+)
+
+go_test(
+    name = "impl_test",
+    srcs = ["jmxlogger_test.go"],
+    embed = [":impl"],
+    gotags = ["test"],
+    deps = [
+        "//comp/core/config",
+        "//comp/def",
+        "//pkg/util/fxutil",
+        "@com_github_stretchr_testify//assert",
+        "@org_uber_go_fx//:fx",
+    ],
+)
