+++ pkg/jmxfetch/jmxfetch.go
@@ -22,7 +22,7 @@ import (
 
 	"go.yaml.in/yaml/v2"
 
-	"github.com/DataDog/datadog-agent/comp/agent/jmxlogger"
+	jmxlogger "github.com/DataDog/datadog-agent/comp/agent/jmxlogger/def"
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery/integration"
 	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
 	dogstatsdConfig "github.com/DataDog/datadog-agent/comp/dogstatsd/config"
